authorizaton
protected resource with authorization

+++ midstates/backend/routes/inventories.js
@@ -88,9 +88,15 @@ router.put(
       description: req.body.description,
     });
     console.log(inventory);
-    Inventory.updateOne({ _id: req.params.id }, inventory).then((result) => {
-      console.log(result);
-      res.status(200).json({ message: "Update successful" });
+    // adding user Id to prevent a different user from deleting the post
+    //TODO add the ability for a super admin to update any posts
+    Inventory.updateOne({ _id: req.params.id, creator: req.userData.userId }, inventory).then((result) => {
+      if (result.nModified > 0) {
+        res.status(200).json({ message: "Update successful" });
+      } else {
+        res.status(401).json({message: "Not authorized"});
+      }
+
     });
   }
 );
@@ -128,10 +134,17 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Inventory.deleteOne({ _id: req.params.id }).then((result) => {
-    console.log(result);
-    res.status(200).json({ message: "Post Deleted!" });
+   // adding user Id to prevent a different user from deleting the post
+    //TODO add the ability for a super admin to delete any posts
+  Inventory.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
+    if (result.n > 0) {
+      res.status(200).json({ message: "Deletion successful" });
+    } else {
+      res.status(401).json({message: "Not authorized to delete this post"});
+    }
+
   });
-});
+}
+);
 
 module.exports = router;
